fastapi-app/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pydantic
gunicorn
